DB error bugfix and minor UI updates
* Fixed: Resolved a WordPress database error caused by a call to a legacy database table.
* Added: Added an empty page placeholder to the "Application Passwords" page for a clearer experience when no items are available.
* Updated: Removed the unwanted border radius from accordion elements in "Manage Plugins" and "Manage Themes."
* Updated: Removed the "Close" button and added a close (X) icon to the "Add Application Password" modal for a cleaner interface.
* Updated: Added a page reload after successfully creating an application password so the new entry appears immediately.

+++ class/class-mainwp-db-site-actions.php
@@ -140,6 +140,9 @@ class MainWP_DB_Site_Actions extends MainWP_DB { // phpcs:ignore Generic.Classes
             return false;
         }
         $table_actions = esc_sql( $this->table_name( 'wp_actions' ) );
+        if ( ! $this->wpdb->get_var( 'SHOW TABLES LIKE ' . $this->wpdb->prepare( '%s', $this->table_name( 'wp_actions' ) ) ) ) {
+            return false;
+        }
         if ( 'action_id' === $by ) {
             if ( $this->wpdb->query( $this->wpdb->prepare( "DELETE FROM {$table_actions} WHERE action_id=%d ", $value ) ) ) {
                 return true;
